# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title
  title: Api
  version: 0.1.0
  description: 드림코어 미로 타이쿤 API
servers:
  - url: /api
    description: API 기본 경로
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: 특정 유저 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 유저 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: 유저 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: 유저 정보 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: 수정된 유저
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: 유저 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 삭제 완료
  /users/{id}/ban:
    post:
      operationId: banUser
      tags: [users]
      summary: 유저 채팅 금지/해제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserBody"
      responses:
        "200":
          description: 처리 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{id}/inventory:
    get:
      operationId: getUserInventory
      tags: [users]
      summary: 유저 인벤토리 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 인벤토리 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: clearUserInventory
      tags: [users]
      summary: 유저 인벤토리 초기화 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 초기화 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ───────────────── GAME ─────────────────
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: 게임 서버 목록
      responses:
        "200":
          description: 서버 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameServer"
    post:
      operationId: createServer
      tags: [servers]
      summary: 서버 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerBody"
      responses:
        "201":
          description: 서버 생성됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
  /servers/{serverId}:
    get:
      operationId: getServer
      tags: [servers]
      summary: 서버 정보 조회
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 서버 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
    patch:
      operationId: updateServer
      tags: [servers]
      summary: 서버 설정 변경
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerBody"
      responses:
        "200":
          description: 수정된 서버
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: 서버 삭제
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 삭제 완료
...
  /shop/buy:
    post:
      operationId: buyShopItem
      tags: [shop]
      summary: 아이템 구매
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyItemBody"
      responses:
        "200":
          description: 구매 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: 잔액 부족 또는 이미 보유
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shop/equip:
    post:
      operationId: equipSkin
      tags: [shop]
      summary: 스킨 장착
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipSkinBody"
      responses:
        "200":
          description: 장착 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ───────────────── LEADERBOARD ─────────────────
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: 신고 목록
      responses:
        "200":
          description: 신고 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: 신고 접수
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: 신고 접수됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{reportId}:
    patch:
      operationId: updateReport
      tags: [reports]
      summary: 신고 상태 변경 (관리자)
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: 변경된 신고
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ───────────────── CHAT ─────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth Schemas ──
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nickname:
          type: string
        role:
          type: string
        currency:
          type: integer
        equippedSkin:
          type: ["string", "null"]
        isBanned:
          type: boolean
        totalScore:
          type: integer
        playtime:
          type: integer
        createdAt:
          type: string
      required: [id, username, nickname, role, currency, isBanned, totalScore, playtime, createdAt]
...
    UpdateUserBody:
      type: object
      properties:
        nickname:
          type: string
        role:
          type: string
        currency:
          type: integer
    BanUserBody:
      type: object
      properties:
        banned:
          type: boolean
      required: [banned]
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: string
        itemType:
          type: string
        acquiredAt:
          type: string
      required: [id, itemId, itemType, acquiredAt]

    # ── Game Schemas ──
...
    GameServer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: integer
        ownerNickname:
          type: string
        mode:
          type: string
        maxPlayers:
          type: integer
        currentPlayers:
          type: integer
        isPublic:
          type: boolean
        complexity:
          type: integer
        trapRate:
          type: number
        mapWarp:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
      required: [id, name, ownerId, mode, maxPlayers, currentPlayers, isPublic, complexity, trapRate, mapWarp, status, createdAt]
    CreateServerBody:
      type: object
      properties:
        name:
          type: string
        mode:
          type: string
        maxPlayers:
          type: integer
        isPublic:
          type: boolean
        complexity:
          type: integer
        trapRate:
          type: number
        mapWarp:
          type: boolean
      required: [name, mode]
    UpdateServerBody:
      type: object
      properties:
        name:
          type: string
        maxPlayers:
          type: integer
        isPublic:
          type: boolean
        complexity:
          type: integer
        trapRate:
          type: number
        mapWarp:
          type: boolean
        status:
          type: string
...
    BuyItemBody:
      type: object
      properties:
        itemId:
          type: string
      required: [itemId]
    EquipSkinBody:
      type: object
      properties:
        itemId:
          type: ["string", "null"]
      required: [itemId]

    # ── Leaderboard ──
...
    Report:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        reporterNickname:
          type: string
        targetNickname:
          type: string
        reason:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, reporterId, targetNickname, reason, status, createdAt]
    CreateReportBody:
      type: object
      properties:
        targetNickname:
          type: string
        reason:
          type: string
      required: [targetNickname, reason]
    UpdateReportBody:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ── Chat ──
...